+++ crates/zng-app/src/widget/node/trace.rs
@@ -9,7 +9,7 @@ use crate::{
     widget::{
         WidgetUpdateMode,
         info::{WidgetInfoBuilder, WidgetLayout, WidgetMeasure},
-        node::{UiNode, UiNodeImpl, UiNodeOpMethod},
+        node::{UiNode, UiNodeImpl, UiNodeMethod},
     },
 };
 
@@ -17,10 +17,10 @@ use super::UiNodeListObserver;
 
 pub(super) struct TraceNode {
     node: UiNode,
-    trace: Box<dyn FnMut(&mut dyn UiNodeImpl, UiNodeOpMethod, &mut dyn FnMut(&mut dyn UiNodeImpl)) + Send + 'static>,
+    trace: Box<dyn FnMut(&mut dyn UiNodeImpl, UiNodeMethod, &mut dyn FnMut(&mut dyn UiNodeImpl)) + Send + 'static>,
 }
 impl TraceNode {
-    pub(super) fn new<S>(node: UiNode, mut enter_mtd: impl FnMut(UiNodeOpMethod) -> S + Send + 'static) -> Self {
+    pub(super) fn new<S>(node: UiNode, mut enter_mtd: impl FnMut(UiNodeMethod) -> S + Send + 'static) -> Self {
         Self {
             node,
             trace: Box::new(move |node, op, call| {
@@ -45,7 +45,6 @@ impl TraceNode {
         }
     }
 }
-// !!: TODO replace UiNodeOpMethod with enum for all methods, with read-only peeking of args
 impl UiNodeImpl for TraceNode {
     fn children_len(&self) -> usize {
         self.node.children_len()
@@ -84,32 +83,34 @@ impl UiNodeImpl for TraceNode {
     }
 
     fn init(&mut self) {
-        (self.trace)(self.node.as_dyn(), UiNodeOpMethod::Init, &mut |n| n.init())
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::Init, &mut |n| n.init())
     }
 
     fn deinit(&mut self) {
-        (self.trace)(self.node.as_dyn(), UiNodeOpMethod::Deinit, &mut |n| n.deinit())
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::Deinit, &mut |n| n.deinit())
     }
 
     fn info(&mut self, info: &mut WidgetInfoBuilder) {
-        (self.trace)(self.node.as_dyn(), UiNodeOpMethod::Info, &mut |n| n.info(info))
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::Info, &mut |n| n.info(info))
     }
 
     fn event(&mut self, update: &crate::update::EventUpdate) {
-        (self.trace)(self.node.as_dyn(), UiNodeOpMethod::Event, &mut |n| n.event(update))
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::Event, &mut |n| n.event(update))
     }
 
     fn update(&mut self, updates: &WidgetUpdates) {
-        (self.trace)(self.node.as_dyn(), UiNodeOpMethod::Update, &mut |n| n.update(updates))
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::Update, &mut |n| n.update(updates))
     }
 
     fn update_list(&mut self, updates: &WidgetUpdates, observer: &mut dyn UiNodeListObserver) {
-        self.node.0.update_list(updates, observer);
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::UpdateList, &mut |n| {
+            n.update_list(updates, observer)
+        })
     }
 
     fn measure(&mut self, wm: &mut WidgetMeasure) -> PxSize {
         let mut out = PxSize::zero();
-        (self.trace)(self.node.as_dyn(), UiNodeOpMethod::Measure, &mut |n| out = n.measure(wm));
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::Measure, &mut |n| out = n.measure(wm));
         out
     }
 
@@ -119,12 +120,16 @@ impl UiNodeImpl for TraceNode {
         measure: &(dyn Fn(usize, &mut UiNode, &mut WidgetMeasure) -> PxSize + Sync),
         fold_size: &(dyn Fn(PxSize, PxSize) -> PxSize + Sync),
     ) -> PxSize {
-        self.node.0.measure_list(wm, measure, fold_size)
+        let mut out = PxSize::zero();
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::MeasureList, &mut |n| {
+            out = n.measure_list(wm, measure, fold_size)
+        });
+        out
     }
 
     fn layout(&mut self, wl: &mut WidgetLayout) -> PxSize {
         let mut out = PxSize::zero();
-        (self.trace)(self.node.as_dyn(), UiNodeOpMethod::Layout, &mut |n| out = n.layout(wl));
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::Layout, &mut |n| out = n.layout(wl));
         out
     }
 
@@ -134,23 +139,29 @@ impl UiNodeImpl for TraceNode {
         layout: &(dyn Fn(usize, &mut UiNode, &mut WidgetLayout) -> PxSize + Sync),
         fold_size: &(dyn Fn(PxSize, PxSize) -> PxSize + Sync),
     ) -> PxSize {
-        self.node.0.layout_list(wl, layout, fold_size)
+        let mut out = PxSize::zero();
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::LayoutList, &mut |n| {
+            out = n.layout_list(wl, layout, fold_size)
+        });
+        out
     }
 
     fn render(&mut self, frame: &mut FrameBuilder) {
-        (self.trace)(self.node.as_dyn(), UiNodeOpMethod::Render, &mut |n| n.render(frame))
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::Render, &mut |n| n.render(frame))
     }
 
     fn render_list(&mut self, frame: &mut FrameBuilder, render: &(dyn Fn(usize, &mut UiNode, &mut FrameBuilder) + Sync)) {
-        self.node.0.render_list(frame, render);
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::RenderList, &mut |n| n.render_list(frame, render))
     }
 
     fn render_update(&mut self, update: &mut FrameUpdate) {
-        (self.trace)(self.node.as_dyn(), UiNodeOpMethod::RenderUpdate, &mut |n| n.render_update(update))
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::RenderUpdate, &mut |n| n.render_update(update))
     }
 
     fn render_update_list(&mut self, update: &mut FrameUpdate, render_update: &(dyn Fn(usize, &mut UiNode, &mut FrameUpdate) + Sync)) {
-        self.node.0.render_update_list(update, render_update);
+        (self.trace)(self.node.as_dyn(), UiNodeMethod::RenderUpdateList, &mut |n| {
+            n.render_update_list(update, render_update)
+        })
     }
 
     fn as_widget(&mut self) -> Option<&mut dyn super::WidgetUiNodeImpl> {