+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This release contains breaking changes that affect the build and app startup.
 * **Breaking** Renamed `ModifiersState` LOGO -> SUPER that was missed from a previous rename.
 * **Breaking** Syntax of the advanced `font_variations!` macro changed to better integrate with `cargo zng fmt`.
 * **Breaking** Advanced Window API crate `zng-ext-window` now has feature `"image"` for optionally compiling image API.
+* **Breaking** Renamed `UiNodeOpMethod` to `UiNodeMethod` and added list methods, to support tracing list nodes.
 
 * Fix feature `zng/config` not enabling required `zng/fs_watcher`.